feat: move enums to migrations
