+++ packages/db/drizzle/0004_jittery_earthquake.sql
@@ -1,4 +1,3 @@
-CREATE TYPE "slack_event_types" AS ENUM ('TEAM_JOIN', 'MESSAGE', 'APP_MENTION', 'MEMBER_JOINED_CHANNEL', 'DND_UPDATED_USER');--> statement-breakpoint
 ALTER TABLE "slack_temporal_raw_events" RENAME COLUMN "text" TO "payload";--> statement-breakpoint
 ALTER TABLE "slack_temporal_raw_events" ADD COLUMN "type" "slack_event_types" NOT NULL;--> statement-breakpoint
 ALTER TABLE "slack_temporal_raw_events" ALTER COLUMN "payload" TYPE jsonb USING payload::jsonb;--> statement-breakpoint
