+++ packages/db/drizzle/0005_silent_leper_queen.sql
@@ -1,4 +1,3 @@
-CREATE TYPE "workspace_deactivation_reason" AS ENUM ('TOKEN_REVOKED', 'APP_UNINSTALLED');--> statement-breakpoint
 ALTER TABLE "workspaces" ADD COLUMN "is_disabled" boolean DEFAULT false NOT NULL;--> statement-breakpoint
 ALTER TABLE "workspaces" ADD COLUMN "deactivation_reason" "workspace_deactivation_reason";--> statement-breakpoint
 ALTER TABLE "workspaces" ADD COLUMN "deactivated_at" timestamp;
