+++ packages/db/drizzle/0009_polite_moonstone.sql
@@ -1,0 +1,1 @@
+CREATE TYPE "public"."workspace_deactivation_reason" AS ENUM('TOKEN_REVOKED', 'APP_UNINSTALLED');
