+++ packages/db/drizzle/meta/_journal.json
@@ -50,6 +50,27 @@
       "when": 1753157777800,
       "tag": "0006_black_marvex",
       "breakpoints": true
+    },
+    {
+      "idx": 7,
+      "version": "7",
+      "when": 1753161637697,
+      "tag": "0007_even_phalanx",
+      "breakpoints": true
+    },
+    {
+      "idx": 8,
+      "version": "7",
+      "when": 1753161939176,
+      "tag": "0008_nappy_young_avengers",
+      "breakpoints": true
+    },
+    {
+      "idx": 9,
+      "version": "7",
+      "when": 1753162069972,
+      "tag": "0009_polite_moonstone",
+      "breakpoints": true
     }
   ]
 }
