+++ packages/db/src/schema/index.ts
@@ -1,4 +1,6 @@
 export * from './dailyFeatures';
+export * from './fatigueScores';
+export * from './interventions';
 export * from './slack/index';
 export * from './user/index';
 export * from './workspace/index';
