+++ packages/db/src/schema/workspace/workspaceExternalProviderTypeEnum.ts
@@ -2,6 +2,6 @@ import { WorkspaceExternalProviderType } from '@calmpulse-app/types';
 import { pgEnum } from 'drizzle-orm/pg-core';
 
 export const workspaceExternalProviderType = pgEnum('workspace_external_provider_type', [
-  WorkspaceExternalProviderType.Slack,
-  WorkspaceExternalProviderType.MicrosoftTeams,
+  WorkspaceExternalProviderType.SLACK,
+  WorkspaceExternalProviderType.MICROSOFT_TEAMS,
 ]);
